+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8001e16d6f8e1484abdeae02a2f981a8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
