+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Explain.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f183c6759748748629e994df1c4e0a09
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
