+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Explain/ExplainCanvas.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 92ca3de516c90440cb2e03bfbf8903b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
