+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/ReadyMenu.cs
@@ -1,0 +1,89 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using TMPro;
+
+/// <summary>
+/// 대기화면 메뉴
+/// 231226 배경택
+/// </summary>
+public class ReadyMenu : MonoBehaviour
+{
+
+    [SerializeField] private GameObject canvases;
+
+    private GameObject gameExplainCanvas;
+    private GameObject settingsCanvas;
+    private GameObject savingCanvas;
+
+
+    private void Awake()
+    {
+        gameExplainCanvas = Utility.FindChildObj(canvases, "ExplainCanvas");
+        settingsCanvas = Utility.FindChildObj(canvases, "SettingCanvas");
+        savingCanvas = Utility.FindChildObj(canvases, "SavingCanvas");
+    }
+
+    // 돌아가기 버튼
+    public void BackToGame()
+    {
+        gameObject.SetActive(false);
+        CanUseSound();
+
+        Controller_Physics.SwitchCameraLock(false);
+    }
+
+    // 도움말 버튼
+    public void GameExplain()
+    {
+        gameExplainCanvas.SetActive(true);
+        gameObject.SetActive(false);
+        CanUseSound();
+
+    }
+
+    // 환경설정창으로 이동
+    public void GoSettings()
+    {
+        settingsCanvas.SetActive(true);
+        gameObject.SetActive(false);
+        CanUseSound();
+    }
+
+    // 게임 저장
+    public void SaveGame()
+    {        
+        DataManager.instance.SaveGameData();        
+        savingCanvas.GetComponent<SavingCanvas>().StartSave();
+        gameObject.SetActive(false);
+    }
+
+    /// <summary>
+    /// 게임 불러오기
+    /// TODO Title쪽으로 옮겨야함
+    /// </summary>
+    public void LoadGame()
+    {
+        //GameEventsManager.instance.dataEvents.LoadData(); // 불러오기 이벤트 발생
+        DataManager.instance.LoadGameData();
+    }
+
+    // 임시 버튼 사용 불가 소리
+    public void CantUseSound()
+    {
+        // 사운드 추가
+        SoundManager.instance.PlaySound(GroupList.UI, (int)UISoundList.Cant_BuySound);
+    }
+
+    // 임시 버튼 사용가능소리
+    private void CanUseSound()
+    {
+        // 사운드 추가       
+        SoundManager.instance.PlaySound(GroupList.UI, (int)UISoundList.Can_BuySound);
+    }
+
+    public void QuitGame()
+    {
+        Application.Quit();
+    }
+}
