+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Save.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cbcf1190992c84d35a459303c636399c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
