+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Save/SavedGamePlayData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1950adb0d33b2470a9d99afdd87cdd08
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
