+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Settings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b5679603f7d584fddbf857e940c311fc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
