+++ Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/ReadyMenu/Settings/VolumeSetting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a63c09fa16cfd4f08997ffc794b8dfdb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
