--- Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/Settings.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8001e16d6f8e1484abdeae02a2f981a8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
