--- Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/Settings/KeyExplainCanvas.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1a74e38e086d247ec8f0f6eea862e893
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
