--- Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/Settings/ReadyMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f5ef78a39dcc04554bb881151eb304db
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
