--- Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/Settings/ResetKeyBindings.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 13c8b72ab152543c09b5a6fb62f214f5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
