--- Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/Settings/Save/MoveObjectData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2d78cb70683f64fb6bf612a56cfa26a3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
