--- Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/Settings/Save/PlayerData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 669b15eea05de4188898007bd14e1a0c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
