--- Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/Settings/Save/SavedSettingsData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 70207043dabb24a54b6a5dbb7c4cf81b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
