--- Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/Settings/Save/SavingCanvas.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e46f3d422cc994a53a8d04b4b01cb243
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
