--- Metalord_btin/MetaLord/Assets/_Test/BKT/Scripts/Settings/VolumeSetting.cs
@@ -1,23 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-/// <summary>
-/// 볼륨 셋팅하는 클래스
-/// 231215 배경택
-/// TODO SettingManager에서 관리할 수 있도록 연계필요
-/// </summary>
-public class VolumeSetting : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
